refactor(tui | config): move Properties struct to config package
